flags file expo.ts renamed to flags.ts
